+++ zcash_primitives/src/sapling/note.rs
@@ -1,10 +1,10 @@
 use group::{ff::Field, GroupEncoding};
 use rand_core::{CryptoRng, RngCore};
+use zcash_spec::PrfExpand;
 
 use super::{
     keys::EphemeralSecretKey, value::NoteValue, Nullifier, NullifierDerivingKey, PaymentAddress,
 };
-use crate::keys::prf_expand;
 
 mod commitment;
 pub use self::commitment::{ExtractedNoteCommitment, NoteCommitment};
@@ -30,7 +30,7 @@ impl Rseed {
         commitment::NoteCommitTrapdoor(match self {
             Rseed::BeforeZip212(rcm) => *rcm,
             Rseed::AfterZip212(rseed) => {
-                jubjub::Fr::from_bytes_wide(prf_expand(rseed, &[0x04]).as_array())
+                jubjub::Fr::from_bytes_wide(&PrfExpand::SAPLING_RCM.with(rseed))
             }
         })
     }
@@ -144,7 +144,7 @@ impl Note {
         match self.rseed {
             Rseed::BeforeZip212(_) => None,
             Rseed::AfterZip212(rseed) => Some(EphemeralSecretKey(jubjub::Fr::from_bytes_wide(
-                prf_expand(&rseed, &[0x05]).as_array(),
+                &PrfExpand::SAPLING_ESK.with(&rseed),
             ))),
         }
     }